generate reasonable latency

+++ generator/topology/service_route.go
@@ -3,6 +3,6 @@ package topology
 type ServiceRoute struct {
 	Route string `json:"route"`
 	DownstreamCalls map[string]string `json:"downstreamCalls,omitempty"`
-	MaxLatencyMillis int `json:"maxLatencyMillis"`
+	MaxLatencyMillis int64 `json:"maxLatencyMillis"`
 	TagSets []TagSet `json:"tagSets"`
 }
